Tidy up/fix tests
